fix: Moved helper functions to lib
