+++ src/app/Home/FriendsList/page.tsx
@@ -1,5 +1,4 @@
 'use client';
-
 import { useEffect, useState } from 'react';
 
 interface Friend {
