+++ src/app/api/getFriends/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import prisma from '@/lib/db';
-import { getUser } from '@/components/getUser';
+import { getUser } from '@/lib/auth/user';
 export async function GET() {
     try {
         const user = await getUser();
