+++ src/app/api/search-user/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import prisma from '@/lib/db';
-import { getUser } from '@/components/getUser';
+import { getUser } from '@/lib/auth/user';
 export async function GET(request: Request) {
   const { searchParams } = new URL(request.url);
   const query = searchParams.get('query');
