+++ src/lib/auth/user.ts
@@ -2,6 +2,7 @@ import { cookies } from 'next/headers';
 import type { User } from '@prisma/client';
 import { verifyJWT } from '@/lib/auth/jwt';
 import prisma from '@/lib/db';
+import { getSession } from "@/lib/auth/session";
 
 export async function getUser(): Promise<User | null> {
   try {
@@ -27,5 +28,33 @@ export async function getUser(): Promise<User | null> {
   } catch (error) {
     console.error('Error getting user:', error);
     return null;
+  }
+}
+
+
+
+export const seedUser = async (token: string | undefined, secret: string):Promise<User|null> => {
+  try {
+    const userSession = await getSession(token, secret);
+    const userId = userSession?.sub
+    if (!userId) {
+      throw new Error('No valid user session found');
+    }
+    const baseURL = process.env.KINDE_SITE_URL
+    const resp = await fetch(`${baseURL}/api/seedUser`,{
+      method: 'POST',
+      headers: {
+        'Content-Type':'application/json'
+      },
+      body: JSON.stringify({userId: userId})
+    });
+    if (!resp.ok){
+      throw new Error ('Failed to seed user')
+    }
+    return await resp.json();
+  }
+  catch (err) {
+   console.log(err)
+   return null
   }
 }
